Begun to add paging options to responseModel

+++ src/helpers/pager.ts
@@ -16,12 +16,13 @@ export function Pager(msg: Message, Client: Client, Response: ResponseModel) {
     
     var m = msg.content.split(' ');
 	let prefix = m.shift();
-    
-    var page = 1;
-    var startPage = parseInt(Response.message);
-    if (startPage != NaN && startPage >= 1 && startPage <= Response.embeds.length) {
-        page = startPage;
+
+    var Options = new PagingOptions();
+    if (Response.pagingoptions != null || Response.pagingoptions != undefined) {
+        Options = Response.pagingoptions;
     }
+
+    var page = Options.startpage.valueOf();
 
     msg.channel.send(Response.embeds[page-1])
         .then(function (message: Message) {
@@ -30,11 +31,17 @@ export function Pager(msg: Message, Client: Client, Response: ResponseModel) {
                 message.react("➡");
             });
 
-            var forwardreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "➡" && user.id == msg.author.id && message.id == reaction.message.id;
-            var backreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "⬅" && user.id == msg.author.id && message.id == reaction.message.id;
+            if (!Options.allowallreactions) {
+                var forwardreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "➡" && user.id == msg.author.id && message.id == reaction.message.id;
+                var backreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "⬅" && user.id == msg.author.id && message.id == reaction.message.id;
+            }
+            else {
+                var forwardreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "➡" && message.id == reaction.message.id;
+                var backreact = (reaction: MessageReaction, user: User) => reaction.emoji.name === "⬅" && message.id == reaction.message.id;
+            }
 
-            var back = message.createReactionCollector(backreact, { time: 600000 });
-            var forward = message.createReactionCollector(forwardreact, { time: 600000 });
+            var back = message.createReactionCollector(backreact, { time: Options.timeout.valueOf() });
+            var forward = message.createReactionCollector(forwardreact, { time: Options.timeout.valueOf() });
 
             back.on('collect', (r) => {
                 page--;
@@ -58,27 +65,41 @@ export function Pager(msg: Message, Client: Client, Response: ResponseModel) {
 
             // trigger 'collect' event on removal of reaction, if valid
             Client.on('messageReactionRemove', (reaction, user) => { 
-                if (!back.ended && back.filter(reaction, user)) { back.emit('collect'); }
-                if (!forward.ended && forward.filter(reaction, user)) { forward.emit('collect'); }
+                if (Options.reactionremoval) {
+                    if (!back.ended && back.filter(reaction, user)) { back.emit('collect'); }
+                    if (!forward.ended && forward.filter(reaction, user)) { forward.emit('collect'); }
+                }
             });
 
             forward.on('end', (s, reason) => {
-                message.delete().then((m) => {
-                    return;
-                }).catch((err)=> {
-                    
-                });
+                if (Options.timeoutdelete) {
+                    message.delete().then((m) => {
+                        return;
+                    }).catch((err)=> {
+                        
+                    });
+                }
             })
 
             back.on('end', (s, reason) => {
-                message.delete().then((m) => {
-                    return;
-                }).catch((err)=> {
+                if (Options.timeoutdelete) {
+                    message.delete().then((m) => {
+                        return;
+                    }).catch((err)=> {
 
-                });
+                    });
+                }
             })
 
         }).catch(function(err) {
             console.log(err);
         });
+}
+
+export class PagingOptions {
+    timeout: Number = 600000;
+    timeoutdelete: Boolean = false;
+    allowallreactions: Boolean = false;
+    reactionremoval: Boolean = true;
+    startpage: Number = 1;
 }
